+++ src/main/java/fm/last/commons/test/file/MethodDataFolder.java
@@ -21,9 +21,10 @@ import org.junit.Test;
 import org.junit.runner.Description;
 import org.junit.runners.model.Statement;
 
-public final class MethodDataFolder extends AbstractDataFolder {
+public final class MethodDataFolder extends DataFolder {
 
   private final File parent;
+  private File folder;
 
   public MethodDataFolder() {
     parent = new File("src" + File.separator + "test" + File.separator + "data");
@@ -46,6 +47,10 @@ public final class MethodDataFolder extends AbstractDataFolder {
         base.evaluate();
       }
     };
+  }
 
+  @Override
+  public File getDataFolder() {
+    return folder;
   }
 }